20250124_BackendUpdate

+++ VisualObjectRecognition/VisualObjectRecognition.Server/Controllers/StorageController.cs
@@ -1,8 +1,9 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
-using VisualObjectRecognition.Server.Data;
+using Microsoft.Azure.Cosmos;
+using System.Text;
 using VisualObjectRecognition.Server.Models;
 using VisualObjectRecognition.Server.Services;
+using static System.Runtime.InteropServices.JavaScript.JSType;
 
 namespace VisualObjectRecognition.Server.Controllers
 {
@@ -12,6 +13,7 @@ namespace VisualObjectRecognition.Server.Controllers
     {
 
         private readonly IStorageRepository _storageRepository;
+
 
         public StorageController(IStorageRepository storageRepository)
         {
@@ -40,13 +42,13 @@ namespace VisualObjectRecognition.Server.Controllers
             try
             {
                 var response = await _storageRepository.GetStorageAsync(id);
-                var user = response.FirstOrDefault();
+                var storage = response.FirstOrDefault();
 
-                if (user == null)
+                if (storage == null)
                 {
                     return NotFound($"Storage with ID {id} not found.");
                 }
-                return Ok(user);
+                return Ok(storage);
             }
             catch (InvalidOperationException ex)
             {
@@ -69,6 +71,7 @@ namespace VisualObjectRecognition.Server.Controllers
 
             try
             {
+
                 var createdStorage = await _storageRepository.CreateStorageAsync(storage);
                 return CreatedAtAction(nameof(GetStorage), new { id = createdStorage.Id }, createdStorage);
             }
@@ -102,18 +105,55 @@ namespace VisualObjectRecognition.Server.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteStorage(string id, string title)
         {
-            try
+            var response = await _storageRepository.GetStorageAsync(id);
+            var storage = response.FirstOrDefault();
+
+            if (storage == null)
             {
-                var success = await _storageRepository.DeleteStorage(id, title);
-                if (success)
-                {
-                    return NoContent();
-                }
                 return NotFound($"Storage with ID {id} not found.");
             }
-            catch (Exception ex)
+            else
             {
-                return StatusCode(500, $"An error occurred: {ex.Message}");
+                //Storage von User entfernen
+                /*if (storage.UserID != null)
+                { 
+                    using (HttpClient client = new HttpClient())
+                    {
+                        try
+                        {
+                            string url = $"https://localhost:7228/api/User/DeleteUserStorage/{storage.UserID}";
+
+                            // JSON-Daten in den HTTP-Content einfügen
+                            var content = new StringContent(storage.Id, Encoding.UTF8, "application/json");
+                            HttpResponseMessage http_response = await client.PutAsync(url, content);
+
+                            if (http_response.IsSuccessStatusCode)
+                            {
+                                string responseContent = await http_response.Content.ReadAsStringAsync();
+                                Console.WriteLine($"Antwort: {responseContent}");
+                            }
+                        }
+                        catch (Exception ex)
+                        {
+                            return BadRequest();
+                        }
+                    }
+                }*/
+
+                //Storage aus DB löschen
+                try
+                {
+                    var success = await _storageRepository.DeleteStorage(id, title);
+                    if (success)
+                    {
+                        return Ok("Löschung erfolgreich!");
+                    }
+                    return NotFound($"Storage with ID {id} not found.");
+                }
+                catch (Exception ex)
+                {
+                    return StatusCode(500, $"An error occurred: {ex.Message}");
+                }
             }
         }
     }
